Add support for SKV signed int key types, and remove support for unsigned types

+++ src/k2/cmd/tpcc/tpcc_client.cpp
@@ -50,7 +50,7 @@ std::vector<String> getRangeEnds(uint32_t numPartitions, uint32_t numWarehouses)
 
     // Warehouse IDs start at 1, and range end is open interval
     for (uint32_t i = 1; i <= numPartitions; ++i) {
-        String range_end = FieldToKeyString<uint32_t>((i*share)+1);
+        String range_end = FieldToKeyString<int16_t>((i*share)+1);
         K2DEBUG("RangeEnd: " << range_end);
         rangeEnds.push_back(range_end);
     }
