+++ src/k2/cmd/tpcc/transactions.h
@@ -338,7 +338,7 @@ private:
     void makeOrderLines() {
         _lines.reserve(*(_order.OrderLineCount));
         _order.AllLocal = 1;
-        for (uint32_t i = 0; i < *(_order.OrderLineCount); ++i) {
+        for (int32_t i = 0; i < *(_order.OrderLineCount); ++i) {
             _lines.emplace_back(_random, _order, i, _max_w_id);
             if (*(_lines.back().SupplyWarehouseID) != _w_id) {
                 _order.AllLocal = 0;
@@ -353,8 +353,8 @@ private:
     RandomContext& _random;
     K23SIClient& _client;
     K2TxnHandle _txn;
-    uint32_t _w_id;
-    uint32_t _max_w_id;
+    int16_t _w_id;
+    int16_t _max_w_id;
     bool _failed;
     Order _order;
     std::vector<OrderLine> _lines;
